requirements.txt
daydream-sdk>=0.1.0
x402-sdk>=0.2.0
ccxt>=3.0.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.23.0
python-dateutil>=2.8.2
requests>=2.28.0
